Untested improvement to F5Calc/F5Collimator    -- This is very ugly code       F5Calc is 99% coverted by Vec3D methods and       Can be a static representation       It is not a tally system because it produces geometry

+++ Model/essBuildInc/BilbaoWheel.h
@@ -62,7 +62,7 @@ class BilbaoWheel : public WheelBase
   double voidRadius;             ///< Final outer radius
   double aspectRatio;            ///< Defines curvature in the yz view
 
-  int nSectors;                  ///< Number of sectors in Tungsten
+  size_t nSectors;                  ///< Number of sectors in Tungsten
   double secSepThick;            ///< Thickness of sector separator
   int secSepMat;                 ///< Material of sector separator
